# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TokyoClinic OS API",
    "description": "REST API contract specification for patient intake, scheduling, consultation, and live queue flows.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/patients": {
      "post": {
        "summary": "Register Patient (Intake)",
        "description": "Performs patient intake and demographic registration. SSN is automatically masked.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "403": {
            "description": "Requires ROLE_RECEPTIONIST or ROLE_ADMIN."
          }
        }
      }
    },
    "/patients/{id}": {
      "get": {
        "summary": "Get Patient Details",
        "description": "Retrieves demographic details. Cached in Redis for 1 hour.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient records retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found."
          }
        }
      },
      "put": {
        "summary": "Update Patient Profile",
        "description": "Updates details and evicts the corresponding Redis patient cache entry.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient profile updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found."
          }
        }
      }
    },
    "/patients/search": {
      "get": {
        "summary": "Search Patients",
        "description": "Performs case-insensitive search by patient name.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results array.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/records": {
      "post": {
        "summary": "Add Medical Record",
        "description": "Appends a consultation record. Custom assessments/vitals are sent in a dynamic JSON structure. Restricted to ROLE_DOCTOR, ROLE_ADMIN.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consultation record saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecord"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Patient Consultation History",
        "description": "Retrieves all consultation records for the patient, ordered by date descending. Restricted to ROLE_DOCTOR, ROLE_ADMIN.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History records array.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/queue/check-in": {
      "post": {
        "summary": "Check-In (Enqueue)",
        "description": "Checks in a patient and issues the next sequential token number. Rejects if patient is already active in queue.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token issued, queued in WAITING state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueItem"
                }
              }
            }
          }
        }
      }
    },
    "/queue/doctor/{doctorId}": {
      "get": {
        "summary": "Get Doctor Queue Lane",
        "description": "Retrieves the active queue list (WAITING and ACTIVE) assigned to the doctor.",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of active queue items.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/queue/{id}/call": {
      "put": {
        "summary": "Call Patient",
        "description": "Sets status to ACTIVE. Dispatches async SMS notifying the patient to proceed to the room. Restricted to ROLE_DOCTOR, ROLE_ADMIN.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient status updated to calling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueItem"
                }
              }
            }
          }
        }
      }
    },
    "/queue/{id}/complete": {
      "put": {
        "summary": "Complete Consultation",
        "description": "Sets status to COMPLETED and evicts patient queue item details. Restricted to ROLE_DOCTOR, ROLE_ADMIN.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consultation finalized, patient removed from active list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueItem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
...
      "PatientRequest": {
        "type": "object",
        "required": ["name", "dateOfBirth", "gender", "ssn", "phone", "email"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "1994-10-27"
          },
          "gender": {
            "type": "string",
            "example": "FEMALE"
          },
          "ssn": {
            "type": "string",
            "example": "999-12-3456"
          },
          "phone": {
            "type": "string",
            "example": "+81-90-1234-5678"
          },
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "bloodType": {
            "type": "string",
            "example": "O_POSITIVE"
          },
          "emergencyContactName": {
            "type": "string",
            "example": "John Doe"
          },
          "emergencyContactPhone": {
            "type": "string",
            "example": "+81-80-8765-4321"
          }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string"
          },
          "ssnMasked": {
            "type": "string",
            "example": "***-**-3456"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "bloodType": {
            "type": "string"
          },
          "emergencyContactName": {
            "type": "string"
          },
          "emergencyContactPhone": {
            "type": "string"
          }
        }
      },
      "MedicalRecordRequest": {
        "type": "object",
        "required": ["conditions", "allergies", "vitals"],
        "properties": {
          "conditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vitals": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "MedicalRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "conditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vitals": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "QueueRequest": {
        "type": "object",
        "required": ["patientId", "doctorId"],
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "QueueItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "tokenNumber": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "checkedInTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}